:recycle: Change the types

+++ src/Liquidator.sol
@@ -42,7 +42,7 @@ contract Liquidator is Ownable {
     struct auctionInformation {
         uint128 openDebt;
         uint128 startBlock;
-        uint8 liqThres;
+        uint16 liqThres;
         uint8 baseCurrency;
         uint128 assetPaid;
         bool stopped;
@@ -119,7 +119,7 @@ contract Liquidator is Ownable {
         address liquidationKeeper,
         address originalOwner,
         uint128 openDebt,
-        uint8 liqThres,
+        uint16 liqThres,
         uint8 baseCurrency
     )
         public
